+++ app/src/main/java/com/example/taskapp/ui/app/FormTaskFragment.kt
@@ -7,21 +7,28 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 
 import com.example.taskapp.databinding.FragmentFormtaskBinding
+import com.example.taskapp.utils.initToolbar
 
 class FormTaskFragment : Fragment() {
-    val _binding: FragmentFormtaskBinding? = null
+    private var _binding: FragmentFormtaskBinding? = null
     private val binding get() = _binding!!
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        val binding = FragmentFormtaskBinding.inflate(inflater, container, false)
+    ): View {
+        _binding = FragmentFormtaskBinding.inflate(inflater, container, false)
         return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        initToolbar(binding.toolbar)
     }
 
     override fun onDestroyView() {
         super.onDestroyView()
+        _binding = null
     }
 
 }