+++ app/src/main/java/com/example/taskapp/ui/auth/RecoveryAccountFragment.kt
@@ -6,6 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.example.taskapp.databinding.FragmentRecoveryAccountBinding
+import com.example.taskapp.utils.initToolbar
 
 class RecoveryAccountFragment : Fragment() {
     private var _binding: FragmentRecoveryAccountBinding? = null
@@ -17,6 +18,11 @@ class RecoveryAccountFragment : Fragment() {
     ): View {
         _binding = FragmentRecoveryAccountBinding.inflate(inflater, container, false)
         return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        initToolbar(binding.toolbar)
     }
 
     override fun onDestroyView() {