+++ CHANGELOG.md
@@ -6,16 +6,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.3.0]
 
 ### Added
-- Toolbar component
 
-[0.2.0]
+- Toolbar funcionality 
+- InitToolbar extension in Fragment 
+- Toolbar component style
+
+## [0.2.0]
 
 ### Added
 - Float button and navigation to FormTask screen
 - FormTask screen styles
 - Tab bar navigation with scroll between tabs
+
 ## [0.1.0]
 
 ### Added
@@ -25,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Login screen styles
 - Splash screen styles
 
+[0.3.0]: https://github.com/tkovs-company/mm-app/releases/tag/v0.3.0
 [0.2.0]: https://github.com/tkovs-company/mm-app/releases/tag/v0.2.0
 [0.1.0]: https://github.com/tkovs-company/mm-app/releases/tag/v0.1.0
 [0.0.0]: https://github.com/tkovs-company/mm-app/releases/tag/v0.0.0